#11346 Add UI Support to Add Custom Profiler Metrics
