+++ openmetadata-ui/src/main/resources/ui/src/components/ProfilerDashboard/profilerDashboard.interface.ts
@@ -55,6 +55,7 @@ export interface ProfilerDetailsCardProps {
   title?: string;
   tickFormatter?: string;
   curveType?: CurveType;
+  isLoading?: boolean;
 }
 
 export enum ProfilerDashboardTab {
