+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/de-de.json
@@ -600,6 +600,7 @@
     "markdown-guide": "Markdown-Leitfaden",
     "market-place": "Marketplace",
     "matches": "Entspricht",
+    "matrix": "Matrix",
     "max": "Maximal",
     "max-login-fail-attampt-plural": "Max login fail attampts",
     "maximum-size-lowercase": "Maximale Größe",
