+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/en-us.json
@@ -600,6 +600,7 @@
     "markdown-guide": "Markdown Guide",
     "market-place": "Marketplace",
     "matches": "Matches",
+    "matrix": "Matrix",
     "max": "Max",
     "max-login-fail-attampt-plural": "Max login fail attampts",
     "maximum-size-lowercase": "maximum size",
