+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/es-es.json
@@ -600,6 +600,7 @@
     "markdown-guide": "Guía de Markdown",
     "market-place": "Marketplace",
     "matches": "Coincide",
+    "matrix": "Matrix",
     "max": "Máx",
     "max-login-fail-attampt-plural": "Max login fail attampts",
     "maximum-size-lowercase": "tamaño máximo",
