+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/fr-fr.json
@@ -600,6 +600,7 @@
     "markdown-guide": "Guide Markdown",
     "market-place": "Marketplace",
     "matches": "Correspond",
+    "matrix": "Matrix",
     "max": "Max",
     "max-login-fail-attampt-plural": "Max login fail attampts",
     "maximum-size-lowercase": "taille maximum",
