+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ja-jp.json
@@ -600,6 +600,7 @@
     "markdown-guide": "Markdownガイド",
     "market-place": "Marketplace",
     "matches": "マッチ",
+    "matrix": "Matrix",
     "max": "最大",
     "max-login-fail-attampt-plural": "Max login fail attampts",
     "maximum-size-lowercase": "最大サイズ",
