+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/pt-br.json
@@ -600,6 +600,7 @@
     "markdown-guide": "Guia Markdown",
     "market-place": "Marketplace",
     "matches": "Combinar",
+    "matrix": "Matrix",
     "max": "Max",
     "max-login-fail-attampt-plural": "Max login fail attampts",
     "maximum-size-lowercase": "tamanho máximo",
