+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/ru-ru.json
@@ -600,6 +600,7 @@
     "markdown-guide": "Markdown Guide",
     "market-place": "Marketplace",
     "matches": "Совпадения",
+    "matrix": "Matrix",
     "max": "Максимум",
     "max-login-fail-attampt-plural": "Max login fail attampts",
     "maximum-size-lowercase": "максимальный размер",
