+++ openmetadata-ui/src/main/resources/ui/src/locale/languages/zh-cn.json
@@ -600,6 +600,7 @@
     "markdown-guide": "Markdown 指南",
     "market-place": "Marketplace",
     "matches": "匹配",
+    "matrix": "Matrix",
     "max": "最大值",
     "max-login-fail-attampt-plural": "Max login fail attampts",
     "maximum-size-lowercase": "最大",
